Edited a lot of things, added settings file, made animation for going to settings and all drinks
